+++ backend/app/project/tests/api/test_configuration_contact_roles.py
@@ -14,9 +14,8 @@ from unittest.mock import patch
 from flask import current_app
 
 from project import base_url
-from project.api.models import Configuration, Contact
+from project.api.models import Configuration, ConfigurationContactRole, Contact, User
 from project.api.models.base_model import db
-from project.api.models.contact_role import ConfigurationContactRole
 from project.extensions.instances import pidinst
 from project.tests.base import BaseTestCase, fake, generate_userinfo_data
 
@@ -283,3 +282,113 @@ class TestConfigurationContactRolesServices(BaseTestCase):
             self.assertEqual(
                 update_external_metadata.call_args.args[0].id, configuration.id
             )
+
+    def test_ensure_unique_constraint_on_post(self):
+        """Ensure that we have a unique constraint for role, configuration and contact."""
+        contact = Contact(
+            given_name="A", family_name="Contact", email="a.contact@localhost"
+        )
+        super_user = User(contact=contact, subject=contact.email, is_superuser=True)
+
+        configuration = Configuration(label="test configuration")
+        role_name = "Owner"
+        role_uri = "https://cv/roles/1"
+
+        contact_role = ConfigurationContactRole(
+            contact=contact,
+            configuration=configuration,
+            role_name=role_name,
+            role_uri=role_uri,
+        )
+
+        db.session.add_all([contact, super_user, configuration, contact_role])
+        db.session.commit()
+
+        payload = {
+            "data": {
+                "type": "configuration_contact_role",
+                "attributes": {
+                    "role_name": role_name,
+                    "role_uri": role_uri,
+                },
+                "relationships": {
+                    "configuration": {
+                        "data": {
+                            "id": configuration.id,
+                            "type": "configuration",
+                        },
+                    },
+                    "contact": {"data": {"id": contact.id, "type": "contact"}},
+                },
+            }
+        }
+
+        with self.run_requests_as(super_user):
+            response = self.client.post(
+                self.url,
+                json=payload,
+                content_type="application/vnd.api+json",
+            )
+        self.assertEqual(response.status_code, 409)
+
+    def test_ensure_unique_constraint_on_patch(self):
+        """Ensure that we have a unique constraint for role, configuration and contact also for changes."""
+        contact = Contact(
+            given_name="A", family_name="Contact", email="a.contact@localhost"
+        )
+        super_user = User(contact=contact, subject=contact.email, is_superuser=True)
+
+        configuration = Configuration(label="test configuration")
+        role_name1 = "Owner"
+        role_uri1 = "https://cv/roles/1"
+
+        contact_role1 = ConfigurationContactRole(
+            contact=contact,
+            configuration=configuration,
+            role_name=role_name1,
+            role_uri=role_uri1,
+        )
+
+        role_name2 = "PI"
+        role_uri2 = "https://cv/roles/2"
+
+        contact_role2 = ConfigurationContactRole(
+            contact=contact,
+            configuration=configuration,
+            role_name=role_name2,
+            role_uri=role_uri2,
+        )
+
+        db.session.add_all(
+            [contact, super_user, configuration, contact_role1, contact_role2]
+        )
+        db.session.commit()
+
+        # It is not possible to add this for the very same configuration, contact & role.
+        payload = {
+            "data": {
+                "type": "configuration_contact_role",
+                "id": contact_role2.id,
+                "attributes": {
+                    "role_name": role_name1,
+                    "role_uri": role_uri1,
+                },
+                "relationships": {
+                    "configuration": {
+                        "data": {
+                            "id": configuration.id,
+                            "type": "configuration",
+                        },
+                    },
+                    "contact": {"data": {"id": contact.id, "type": "contact"}},
+                },
+            }
+        }
+
+        with self.run_requests_as(super_user):
+            response = self.client.patch(
+                f"{self.url}/{contact_role2.id}",
+                json=payload,
+                content_type="application/vnd.api+json",
+            )
+        self.assertEqual(response.status_code, 409)